feat: validate blueprint config when executing blueprints
